+++ src/ZiggyServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace Tightenco\Ziggy;
 
-use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\View\Compilers\BladeCompiler;
 
@@ -10,14 +9,6 @@ class ZiggyServiceProvider extends ServiceProvider
 {
     public function boot()
     {
-        Route::macro('except', function ($group = null) {
-            return Macro::except($this, $group);
-        });
-
-        Route::macro('only', function ($group = null) {
-            return Macro::only($this, $group);
-        });
-
         if ($this->app->resolved('blade.compiler')) {
             $this->registerDirective($this->app['blade.compiler']);
         } else {
@@ -27,9 +18,7 @@ class ZiggyServiceProvider extends ServiceProvider
         }
 
         if ($this->app->runningInConsole()) {
-            $this->commands([
-                CommandRouteGenerator::class,
-            ]);
+            $this->commands(CommandRouteGenerator::class);
         }
     }