+++ tests/Unit/ZiggyTest.php
@@ -2,10 +2,10 @@
 
 namespace Tests\Unit;
 
-use Tests\TestCase;
-use Tightenco\Ziggy\RoutePayload;
+use Tightenco\Tests\TestCase;
+use Tightenco\Ziggy\Ziggy;
 
-class RoutePayloadTest extends TestCase
+class ZiggyTest extends TestCase
 {
     protected $router;
 
@@ -22,13 +22,19 @@ class RoutePayloadTest extends TestCase
         $this->router->post('posts', $this->noop())->middleware(['auth', 'role:admin'])->name('posts.store');
         $this->router->get('admin/users', $this->noop())->middleware(['role:admin'])->name('admin.users.index');
 
+        if ($this->laravelVersion(7)) {
+            $this->router->get('/posts/{post}/comments/{comment:uuid}', function () {
+                return '';
+            })->name('postComments.show');
+        }
+
         $this->router->getRoutes()->refreshNameLookups();
     }
 
     /** @test */
     public function only_matching_routes_included_with_include_enabled()
     {
-        $routePayload = new RoutePayload($this->router);
+        $routePayload = new Ziggy;
         $filters = ['posts.s*', 'home'];
         $routes = $routePayload->filter($filters, true);
 
@@ -50,13 +56,19 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+        }
+
         $this->assertEquals($expected, $routes->toArray());
     }
 
     /** @test */
     public function only_matching_routes_excluded_with_exclude_enabled()
     {
-        $routePayload = new RoutePayload($this->router);
+        $routePayload = new Ziggy;
         $filters = ['posts.s*', 'home', 'admin.*'];
         $routes = $routePayload->filter($filters, false);
 
@@ -73,6 +85,21 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+
+            $expected['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
         $this->assertEquals($expected, $routes->toArray());
     }
 
@@ -83,7 +110,7 @@ class RoutePayloadTest extends TestCase
             'only' => ['posts.s*', 'home'],
         ]);
 
-        $routes = RoutePayload::compile($this->router);
+        $routes = (new Ziggy)->toArray()['namedRoutes'];
 
         $expected = [
             'home' => [
@@ -103,7 +130,13 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $routes->toArray());
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+        }
+
+        $this->assertEquals($expected, $routes);
     }
 
     /** @test */
@@ -113,7 +146,7 @@ class RoutePayloadTest extends TestCase
             'except' => ['posts.s*', 'home', 'admin.*'],
         ]);
 
-        $routes = RoutePayload::compile($this->router);
+        $routes = (new Ziggy)->toArray()['namedRoutes'];
 
         $expected = [
             'posts.index' => [
@@ -128,7 +161,22 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $routes->toArray());
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+
+            $expected['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
+        $this->assertEquals($expected, $routes);
     }
 
     /** @test */
@@ -139,7 +187,7 @@ class RoutePayloadTest extends TestCase
             'only' => ['home'],
         ]);
 
-        $routes = RoutePayload::compile($this->router);
+        $routes = (new Ziggy)->toArray()['namedRoutes'];
 
         $expected = [
             'home' => [
@@ -174,7 +222,22 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $routes->toArray());
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+
+            $expected['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
+        $this->assertEquals($expected, $routes);
     }
 
     /** @test */
@@ -186,7 +249,7 @@ class RoutePayloadTest extends TestCase
             ],
         ]);
 
-        $routes = RoutePayload::compile($this->router, 'authors');
+        $routes = (new Ziggy('authors'))->toArray()['namedRoutes'];
 
         $expected = [
             'home' => [
@@ -211,14 +274,20 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $routes->toArray());
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+        }
+
+        $this->assertEquals($expected, $routes);
     }
 
     // can_compile_route_payload
     /** @test */
     public function non_existence_of_group_returns_unfiltered_routes()
     {
-        $routes = RoutePayload::compile($this->router, 'authors');
+        $routes = (new Ziggy('authors'))->toArray()['namedRoutes'];
 
         $expected = [
             'home' => [
@@ -253,7 +322,22 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $routes->toArray());
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+
+            $expected['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
+        $this->assertEquals($expected, $routes);
     }
 
     /** @test */
@@ -263,7 +347,7 @@ class RoutePayloadTest extends TestCase
             'middleware' => true,
         ]);
 
-        $routes = RoutePayload::compile($this->router);
+        $routes = (new Ziggy)->toArray()['namedRoutes'];
 
         $expected = [
             'home' => [
@@ -304,7 +388,23 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $routes->toArray());
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+
+            $expected['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'middleware' => [],
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
+        $this->assertEquals($expected, $routes);
     }
 
     /** @test */
@@ -314,7 +414,7 @@ class RoutePayloadTest extends TestCase
             'middleware' => ['auth'],
         ]);
 
-        $routes = RoutePayload::compile($this->router);
+        $routes = (new Ziggy)->toArray()['namedRoutes'];
 
         $expected = [
             'home' => [
@@ -355,6 +455,141 @@ class RoutePayloadTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $routes->toArray());
+        if ($this->laravelVersion(7)) {
+            foreach ($expected as $key => $route) {
+                $expected[$key]['bindings'] = [];
+            }
+
+            $expected['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'middleware' => [],
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
+        $this->assertEquals($expected, $routes);
+    }
+
+    /** @test */
+    public function route_payload_can_array_itself()
+    {
+        $ziggy = new Ziggy;
+
+        $expected = [
+            'baseUrl' => 'http://myapp.com/',
+            'baseProtocol' => 'http',
+            'baseDomain' => 'myapp.com',
+            'basePort' => null,
+            'defaultParameters' => [],
+            'namedRoutes' => [
+                'home' => [
+                    'uri' => 'home',
+                    'methods' => ['GET', 'HEAD'],
+                    'domain' => null,
+                ],
+                'posts.index' => [
+                    'uri' => 'posts',
+                    'methods' => ['GET', 'HEAD'],
+                    'domain' => null,
+                ],
+                'posts.show' => [
+                    'uri' => 'posts/{post}',
+                    'methods' => ['GET', 'HEAD'],
+                    'domain' => null,
+                ],
+                'postComments.index' => [
+                    'uri' => 'posts/{post}/comments',
+                    'methods' => ['GET', 'HEAD'],
+                    'domain' => null,
+                ],
+                'posts.store' => [
+                    'uri' => 'posts',
+                    'methods' => ['POST'],
+                    'domain' => null,
+                ],
+                'admin.users.index' => [
+                    'uri' => 'admin/users',
+                    'methods' => ['GET', 'HEAD'],
+                    'domain' => null,
+                ],
+            ],
+        ];
+
+        if ($this->laravelVersion(7)) {
+            foreach ($expected['namedRoutes'] as $key => $route) {
+                $expected['namedRoutes'][$key]['bindings'] = [];
+            }
+
+            $expected['namedRoutes']['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
+        $this->assertSame($expected, $ziggy->toArray());
+        $this->assertSame($expected, $ziggy->jsonSerialize());
+    }
+
+    /** @test */
+    public function route_payload_can_json_itself()
+    {
+        app('config')->set('ziggy', ['only' => ['postComments.*']]);
+        $ziggy = new Ziggy;
+
+        $expected = [
+            'baseUrl' => 'http://myapp.com/',
+            'baseProtocol' => 'http',
+            'baseDomain' => 'myapp.com',
+            'basePort' => null,
+            'defaultParameters' => [],
+            'namedRoutes' => [
+                'postComments.index' => [
+                    'uri' => 'posts/{post}/comments',
+                    'methods' => ['GET', 'HEAD'],
+                    'domain' => null,
+                ],
+            ],
+        ];
+
+        if ($this->laravelVersion(7)) {
+            $expected['namedRoutes']['postComments.index']['bindings'] = [];
+
+            $expected['namedRoutes']['postComments.show'] = [
+                'uri' => 'posts/{post}/comments/{comment}',
+                'methods' => ['GET', 'HEAD'],
+                'domain' => null,
+                'bindings' => [
+                    'comment' => 'uuid',
+                ],
+            ];
+        }
+
+        $json = '{"baseUrl":"http:\/\/myapp.com\/","baseProtocol":"http","baseDomain":"myapp.com","basePort":null,"defaultParameters":[],"namedRoutes":{"postComments.index":{"uri":"posts\/{post}\/comments","methods":["GET","HEAD"],"domain":null' . ($this->laravelVersion(7) ? ',"bindings":[]},"postComments.show":{"uri":"posts\/{post}\/comments\/{comment}","methods":["GET","HEAD"],"domain":null,"bindings":{"comment":"uuid"}' : '') . '}}}';
+
+        $this->assertSame($expected, json_decode(json_encode($ziggy), true));
+        $this->assertSame($json, json_encode($ziggy));
+        $this->assertSame($json, $ziggy->toJson());
+    }
+
+    /** @test */
+    public function route_payload_can_automatically_json_itself_as_part_of_a_response()
+    {
+        app('config')->set('ziggy', ['only' => ['postComments.*']]);
+        $this->router->get('json', function () {
+            return response()->json(new Ziggy);
+        });
+
+        $response = $this->get('json');
+
+        $response->assertSuccessful();
+        $this->assertSame((new Ziggy)->toJson(), $response->getContent());
     }
 }